refactor: OnboardingPageを4つのステップコンポーネントに分割
474行のOnboardingPageをステップ単位で独立コンポーネントに分割:
- OnboardingProfileStep: プロフィール入力（名前・自己紹介）
- OnboardingSkillsStep: スキル選択（言語・フレームワーク）
- OnboardingIntegrationsStep: 外部サービス連携（GitHub/Zenn/Qiita/AtCoder/paiza）
- OnboardingCompleteStep: 完了画面

OnboardingPageは130行に縮小し、ステップインジケーターとルーティングに専念。
